// test-task/src/app/dashboard/constants/tooltip-data.enum.ts
export enum TooltipMessages  {
    error = 'Сообщение об ошибке',
    warning = 'Предупредительное сообщение',
    success = 'Сообщение об успехе'
};

export enum TooltipColor {
    red = 'red',
    gold = 'gold',
    green = 'green',
}

export enum TooltipButtonText {
    error = 'Ошибка',
    warning = 'Предупреждение',
    success = 'Успех',
    message = 'Сообщение',
}